refactor: Clean up chart config fixtures
