--- app/test/__fixtures/config/dev/QGYSf81ILfWG.json
@@ -1,32 +1,0 @@
-{
-  "key": "QGYSf81ILfWG",
-  "data": {
-    "meta": {
-      "title": { "de": "", "en": "", "fr": "", "it": "" },
-      "description": { "de": "", "en": "", "fr": "", "it": "" }
-    },
-    "dataSet": "http://environment.ld.admin.ch/foen/rl_trial/dataset",
-    "chartConfig": {
-      "version": "1.2.1",
-      "fields": {
-        "x": {
-          "sorting": {
-            "sortingType": "byDimensionLabel",
-            "sortingOrder": "asc"
-          },
-          "componentIri": "http://environment.ld.admin.ch/foen/rl_trial/dimension/0"
-        },
-        "y": {
-          "componentIri": "http://environment.ld.admin.ch/foen/rl_trial/measure/RE"
-        }
-      },
-      "filters": {
-        "http://environment.ld.admin.ch/foen/rl_trial/dimension/1": {
-          "type": "single",
-          "value": "http://environment.ld.admin.ch/foen/rl_trial/artenGruppe/MAMM"
-        }
-      },
-      "chartType": "column"
-    }
-  }
-}
