--- app/test/__fixtures/config/dev/QzgH15yNMeuZ.json
@@ -1,40 +1,0 @@
-{
-  "key": "QzgH15yNMeuZ",
-  "data": {
-    "meta": {
-      "title": { "de": "", "en": "", "fr": "", "it": "" },
-      "description": { "de": "", "en": "", "fr": "", "it": "" }
-    },
-    "dataSet": "http://environment.ld.admin.ch/foen/px/0703010000_103/dataset",
-    "chartConfig": {
-      "version": "1.2.1",
-      "fields": {
-        "x": {
-          "sorting": {
-            "sortingType": "byDimensionLabel",
-            "sortingOrder": "asc"
-          },
-          "componentIri": "http://environment.ld.admin.ch/foen/px/0703010000_103/dimension/0"
-        },
-        "y": {
-          "componentIri": "http://environment.ld.admin.ch/foen/px/0703010000_103/measure/0"
-        }
-      },
-      "filters": {
-        "http://environment.ld.admin.ch/foen/px/0703010000_103/dimension/1": {
-          "type": "single",
-          "value": "http://environment.ld.admin.ch/foen/px/0703010000_103/dimension/1/0"
-        },
-        "http://environment.ld.admin.ch/foen/px/0703010000_103/dimension/2": {
-          "type": "single",
-          "value": "http://environment.ld.admin.ch/foen/px/0703010000_103/dimension/2/0"
-        },
-        "http://environment.ld.admin.ch/foen/px/0703010000_103/dimension/3": {
-          "type": "single",
-          "value": "http://environment.ld.admin.ch/foen/px/0703010000_103/dimension/3/0"
-        }
-      },
-      "chartType": "column"
-    }
-  }
-}
