--- app/test/__fixtures/config/dev/UElM_bRD-FNL.json
@@ -1,74 +1,0 @@
-{
-  "key": "UElM_bRD-FNL",
-  "data": {
-    "meta": {
-      "title": { "de": "", "en": "", "fr": "", "it": "" },
-      "description": { "de": "", "en": "", "fr": "", "it": "" }
-    },
-    "dataSet": "http://environment.ld.admin.ch/foen/px/0703010000_106/dataset",
-    "chartConfig": {
-      "version": "1.2.1",
-      "fields": {
-        "x": {
-          "componentIri": "http://environment.ld.admin.ch/foen/px/0703010000_106/dimension/0"
-        },
-        "y": {
-          "componentIri": "http://environment.ld.admin.ch/foen/px/0703010000_106/measure/1"
-        },
-        "segment": {
-          "type": "stacked",
-          "palette": "category10",
-          "sorting": {
-            "sortingType": "byDimensionLabel",
-            "sortingOrder": "asc"
-          },
-          "colorMapping": {
-            "http://environment.ld.admin.ch/foen/px/0703010000_106/dimension/2/0": "#1f77b4",
-            "http://environment.ld.admin.ch/foen/px/0703010000_106/dimension/2/1": "#ff7f0e",
-            "http://environment.ld.admin.ch/foen/px/0703010000_106/dimension/2/2": "#2ca02c",
-            "http://environment.ld.admin.ch/foen/px/0703010000_106/dimension/2/3": "#d62728",
-            "http://environment.ld.admin.ch/foen/px/0703010000_106/dimension/2/4": "#9467bd",
-            "http://environment.ld.admin.ch/foen/px/0703010000_106/dimension/2/5": "#8c564b",
-            "http://environment.ld.admin.ch/foen/px/0703010000_106/dimension/2/6": "#e377c2",
-            "http://environment.ld.admin.ch/foen/px/0703010000_106/dimension/2/7": "#7f7f7f",
-            "http://environment.ld.admin.ch/foen/px/0703010000_106/dimension/2/8": "#bcbd22",
-            "http://environment.ld.admin.ch/foen/px/0703010000_106/dimension/2/9": "#17becf",
-            "http://environment.ld.admin.ch/foen/px/0703010000_106/dimension/2/10": "#1f77b4",
-            "http://environment.ld.admin.ch/foen/px/0703010000_106/dimension/2/11": "#ff7f0e",
-            "http://environment.ld.admin.ch/foen/px/0703010000_106/dimension/2/12": "#2ca02c",
-            "http://environment.ld.admin.ch/foen/px/0703010000_106/dimension/2/13": "#d62728",
-            "http://environment.ld.admin.ch/foen/px/0703010000_106/dimension/2/14": "#9467bd",
-            "http://environment.ld.admin.ch/foen/px/0703010000_106/dimension/2/15": "#8c564b",
-            "http://environment.ld.admin.ch/foen/px/0703010000_106/dimension/2/16": "#e377c2",
-            "http://environment.ld.admin.ch/foen/px/0703010000_106/dimension/2/17": "#7f7f7f",
-            "http://environment.ld.admin.ch/foen/px/0703010000_106/dimension/2/18": "#bcbd22",
-            "http://environment.ld.admin.ch/foen/px/0703010000_106/dimension/2/19": "#17becf",
-            "http://environment.ld.admin.ch/foen/px/0703010000_106/dimension/2/20": "#1f77b4",
-            "http://environment.ld.admin.ch/foen/px/0703010000_106/dimension/2/21": "#ff7f0e",
-            "http://environment.ld.admin.ch/foen/px/0703010000_106/dimension/2/22": "#2ca02c",
-            "http://environment.ld.admin.ch/foen/px/0703010000_106/dimension/2/23": "#d62728",
-            "http://environment.ld.admin.ch/foen/px/0703010000_106/dimension/2/24": "#9467bd",
-            "http://environment.ld.admin.ch/foen/px/0703010000_106/dimension/2/25": "#8c564b",
-            "http://environment.ld.admin.ch/foen/px/0703010000_106/dimension/2/26": "#e377c2"
-          },
-          "componentIri": "http://environment.ld.admin.ch/foen/px/0703010000_106/dimension/2"
-        }
-      },
-      "filters": {
-        "http://environment.ld.admin.ch/foen/px/0703010000_106/dimension/1": {
-          "type": "single",
-          "value": "http://environment.ld.admin.ch/foen/px/0703010000_106/dimension/1/0"
-        },
-        "http://environment.ld.admin.ch/foen/px/0703010000_106/dimension/3": {
-          "type": "single",
-          "value": "http://environment.ld.admin.ch/foen/px/0703010000_106/dimension/3/0"
-        },
-        "http://environment.ld.admin.ch/foen/px/0703010000_106/dimension/4": {
-          "type": "single",
-          "value": "http://environment.ld.admin.ch/foen/px/0703010000_106/dimension/4/0"
-        }
-      },
-      "chartType": "area"
-    }
-  }
-}
