--- app/test/__fixtures/config/dev/cFd4UuIwNd8E.json
@@ -1,36 +1,0 @@
-{
-  "key": "cFd4UuIwNd8E",
-  "data": {
-    "meta": {
-      "title": { "de": "", "en": "", "fr": "", "it": "" },
-      "description": { "de": "", "en": "", "fr": "", "it": "" }
-    },
-    "dataSet": "http://environment.ld.admin.ch/foen/px/0703030000_134/dataset",
-    "chartConfig": {
-      "version": "1.2.1",
-      "fields": {
-        "x": {
-          "sorting": {
-            "sortingType": "byDimensionLabel",
-            "sortingOrder": "asc"
-          },
-          "componentIri": "http://environment.ld.admin.ch/foen/px/0703030000_134/dimension/0"
-        },
-        "y": {
-          "componentIri": "http://environment.ld.admin.ch/foen/px/0703030000_134/measure/0"
-        }
-      },
-      "filters": {
-        "http://environment.ld.admin.ch/foen/px/0703030000_134/dimension/1": {
-          "type": "single",
-          "value": "http://environment.ld.admin.ch/foen/px/0703030000_134/dimension/1/0"
-        },
-        "http://environment.ld.admin.ch/foen/px/0703030000_134/dimension/2": {
-          "type": "single",
-          "value": "http://environment.ld.admin.ch/foen/px/0703030000_134/dimension/2/0"
-        }
-      },
-      "chartType": "column"
-    }
-  }
-}
