--- app/test/__fixtures/config/dev/cYUA63lm2Avv.json
@@ -1,70 +1,0 @@
-{
-  "key": "cYUA63lm2Avv",
-  "data": {
-    "meta": {
-      "title": { "de": "", "en": "", "fr": "", "it": "" },
-      "description": { "de": "", "en": "", "fr": "", "it": "" }
-    },
-    "dataSet": "http://environment.ld.admin.ch/foen/px/0703030000_112/dataset",
-    "chartConfig": {
-      "version": "1.2.1",
-      "fields": {
-        "x": {
-          "componentIri": "http://environment.ld.admin.ch/foen/px/0703030000_112/dimension/2"
-        },
-        "y": {
-          "componentIri": "http://environment.ld.admin.ch/foen/px/0703030000_112/measure/0"
-        }
-      },
-      "filters": {
-        "http://environment.ld.admin.ch/foen/px/0703030000_112/dimension/0": {
-          "type": "single",
-          "value": "http://environment.ld.admin.ch/foen/px/0703030000_112/dimension/0/0"
-        },
-        "http://environment.ld.admin.ch/foen/px/0703030000_112/dimension/1": {
-          "type": "single",
-          "value": "http://environment.ld.admin.ch/foen/px/0703030000_112/dimension/1/0"
-        },
-        "http://environment.ld.admin.ch/foen/px/0703030000_112/dimension/2": {
-          "type": "multi",
-          "values": {
-            "http://environment.ld.admin.ch/foen/px/0703030000_112/dimension/2/0": true,
-            "http://environment.ld.admin.ch/foen/px/0703030000_112/dimension/2/3": true,
-            "http://environment.ld.admin.ch/foen/px/0703030000_112/dimension/2/4": true,
-            "http://environment.ld.admin.ch/foen/px/0703030000_112/dimension/2/5": true,
-            "http://environment.ld.admin.ch/foen/px/0703030000_112/dimension/2/6": true,
-            "http://environment.ld.admin.ch/foen/px/0703030000_112/dimension/2/7": true,
-            "http://environment.ld.admin.ch/foen/px/0703030000_112/dimension/2/8": true,
-            "http://environment.ld.admin.ch/foen/px/0703030000_112/dimension/2/9": true,
-            "http://environment.ld.admin.ch/foen/px/0703030000_112/dimension/2/10": true,
-            "http://environment.ld.admin.ch/foen/px/0703030000_112/dimension/2/11": true,
-            "http://environment.ld.admin.ch/foen/px/0703030000_112/dimension/2/12": true,
-            "http://environment.ld.admin.ch/foen/px/0703030000_112/dimension/2/13": true,
-            "http://environment.ld.admin.ch/foen/px/0703030000_112/dimension/2/14": true,
-            "http://environment.ld.admin.ch/foen/px/0703030000_112/dimension/2/15": true,
-            "http://environment.ld.admin.ch/foen/px/0703030000_112/dimension/2/16": true,
-            "http://environment.ld.admin.ch/foen/px/0703030000_112/dimension/2/17": true,
-            "http://environment.ld.admin.ch/foen/px/0703030000_112/dimension/2/18": true,
-            "http://environment.ld.admin.ch/foen/px/0703030000_112/dimension/2/19": true,
-            "http://environment.ld.admin.ch/foen/px/0703030000_112/dimension/2/20": true,
-            "http://environment.ld.admin.ch/foen/px/0703030000_112/dimension/2/21": true,
-            "http://environment.ld.admin.ch/foen/px/0703030000_112/dimension/2/22": true,
-            "http://environment.ld.admin.ch/foen/px/0703030000_112/dimension/2/23": true,
-            "http://environment.ld.admin.ch/foen/px/0703030000_112/dimension/2/24": true,
-            "http://environment.ld.admin.ch/foen/px/0703030000_112/dimension/2/25": true,
-            "http://environment.ld.admin.ch/foen/px/0703030000_112/dimension/2/26": true
-          }
-        },
-        "http://environment.ld.admin.ch/foen/px/0703030000_112/dimension/3": {
-          "type": "single",
-          "value": "http://environment.ld.admin.ch/foen/px/0703030000_112/dimension/3/0"
-        },
-        "http://environment.ld.admin.ch/foen/px/0703030000_112/dimension/4": {
-          "type": "single",
-          "value": "http://environment.ld.admin.ch/foen/px/0703030000_112/dimension/4/0"
-        }
-      },
-      "chartType": "column"
-    }
-  }
-}
