--- app/test/__fixtures/config/dev/u7SwBlMbyT2v.json
@@ -1,44 +1,0 @@
-{
-  "key": "u7SwBlMbyT2v",
-  "data": {
-    "meta": {
-      "title": { "de": "Hhhh", "en": "Hhh", "fr": "Vvb", "it": "Vbbj" },
-      "description": { "de": "", "en": "", "fr": "", "it": "" }
-    },
-    "dataSet": "http://environment.ld.admin.ch/foen/px/0703030000_112/dataset",
-    "chartConfig": {
-      "version": "1.2.1",
-      "fields": {
-        "x": {
-          "sorting": {
-            "sortingType": "byDimensionLabel",
-            "sortingOrder": "asc"
-          },
-          "componentIri": "http://environment.ld.admin.ch/foen/px/0703030000_112/dimension/0"
-        },
-        "y": {
-          "componentIri": "http://environment.ld.admin.ch/foen/px/0703030000_112/measure/0"
-        }
-      },
-      "filters": {
-        "http://environment.ld.admin.ch/foen/px/0703030000_112/dimension/1": {
-          "type": "single",
-          "value": "http://environment.ld.admin.ch/foen/px/0703030000_112/dimension/1/0"
-        },
-        "http://environment.ld.admin.ch/foen/px/0703030000_112/dimension/2": {
-          "type": "single",
-          "value": "http://environment.ld.admin.ch/foen/px/0703030000_112/dimension/2/0"
-        },
-        "http://environment.ld.admin.ch/foen/px/0703030000_112/dimension/3": {
-          "type": "single",
-          "value": "http://environment.ld.admin.ch/foen/px/0703030000_112/dimension/3/0"
-        },
-        "http://environment.ld.admin.ch/foen/px/0703030000_112/dimension/4": {
-          "type": "single",
-          "value": "http://environment.ld.admin.ch/foen/px/0703030000_112/dimension/4/0"
-        }
-      },
-      "chartType": "column"
-    }
-  }
-}
