--- app/test/__fixtures/config/int/column-red-list.json
@@ -1,79 +1,0 @@
-{
-  "key": "nQckT-nFEQe4",
-  "data": {
-    "meta": {
-      "title": {
-        "de": "",
-        "en": "",
-        "fr": "",
-        "it": ""
-      },
-      "description": {
-        "de": "",
-        "en": "",
-        "fr": "",
-        "it": ""
-      }
-    },
-    "dataSet": "https://environment.ld.admin.ch/foen/UBD003002/5",
-    "dataSource": {
-      "type": "sparql",
-      "url": "https://lindas-cached.int.cluster.ldbar.ch/query"
-    },
-    "chartConfig": {
-      "version": "1.2.1",
-      "fields": {
-        "x": {
-          "sorting": {
-            "sortingType": "byDimensionLabel",
-            "sortingOrder": "asc"
-          },
-          "componentIri": "https://environment.ld.admin.ch/foen/UBD003002/artengruppe"
-        },
-        "y": {
-          "componentIri": "https://environment.ld.admin.ch/foen/UBD003002/value"
-        },
-        "segment": {
-          "type": "stacked",
-          "palette": "category10",
-          "sorting": {
-            "sortingType": "byDimensionLabel",
-            "sortingOrder": "asc"
-          },
-          "colorMapping": {
-            "https://environment.ld.admin.ch/foen/UBD003002/Status/CR": "#8c564b",
-            "https://environment.ld.admin.ch/foen/UBD003002/Status/DD": "#1f77b4",
-            "https://environment.ld.admin.ch/foen/UBD003002/Status/EN": "#9467bd",
-            "https://environment.ld.admin.ch/foen/UBD003002/Status/EX": "#7f7f7f",
-            "https://environment.ld.admin.ch/foen/UBD003002/Status/LC": "#ff7f0e",
-            "https://environment.ld.admin.ch/foen/UBD003002/Status/NT": "#2ca02c",
-            "https://environment.ld.admin.ch/foen/UBD003002/Status/RE": "#e377c2",
-            "https://environment.ld.admin.ch/foen/UBD003002/Status/VU": "#d62728"
-          },
-          "componentIri": "https://environment.ld.admin.ch/foen/UBD003002/status"
-        }
-      },
-      "filters": {},
-      "chartType": "column",
-      "interactiveFiltersConfig": {
-        "timeRange": {
-          "active": false,
-          "presets": {
-            "to": "",
-            "from": "",
-            "type": "range"
-          },
-          "componentIri": ""
-        },
-        "legend": {
-          "active": false,
-          "componentIri": ""
-        },
-        "dataFilters": {
-          "active": false,
-          "componentIris": []
-        }
-      }
-    }
-  }
-}
