--- app/test/__fixtures/config/int/column-waldflasche.json
@@ -1,55 +1,0 @@
-{
-  "key": "clQmsblRxpK7",
-  "data": {
-    "state": "CONFIGURING_CHART",
-    "dataSet": "https://environment.ld.admin.ch/foen/nfi/49-19-None-None-44/cube/1",
-    "dataSource": {
-      "type": "sparql",
-      "url": "https://lindas-cached.int.cluster.ldbar.ch/query"
-    },
-    "meta": {
-      "title": { "de": "", "fr": "", "it": "", "en": "" },
-      "description": { "de": "", "fr": "", "it": "", "en": "" }
-    },
-    "chartConfig": {
-      "version": "1.2.1",
-      "chartType": "column",
-      "filters": {
-        "https://environment.ld.admin.ch/foen/nfi/inventory": {
-          "type": "single",
-          "value": "https://environment.ld.admin.ch/foen/nfi/Inventory/150"
-        },
-        "https://environment.ld.admin.ch/foen/nfi/struk": {
-          "type": "single",
-          "value": "https://environment.ld.admin.ch/foen/nfi/ClassificationUnit/Struk/-1"
-        },
-        "https://environment.ld.admin.ch/foen/nfi/prodreg": {
-          "type": "single",
-          "value": "https://environment.ld.admin.ch/foen/nfi/UnitOfReference/Prodreg/1"
-        },
-        "https://environment.ld.admin.ch/foen/nfi/grid": {
-          "type": "single",
-          "value": "https://environment.ld.admin.ch/foen/nfi/Grid/N4"
-        }
-      },
-      "interactiveFiltersConfig": {
-        "legend": { "active": false, "componentIri": "" },
-        "timeRange": {
-          "active": false,
-          "componentIri": "",
-          "presets": { "type": "range", "from": "", "to": "" }
-        },
-        "dataFilters": { "active": false, "componentIris": [] }
-      },
-      "fields": {
-        "x": {
-          "componentIri": "https://environment.ld.admin.ch/foen/nfi/unitOfEvaluation"
-        },
-        "y": {
-          "componentIri": "https://environment.ld.admin.ch/foen/nfi/forestArea"
-        }
-      }
-    },
-    "activeField": "x"
-  }
-}
