+++ app/test/__fixtures/config/int/configs.ts
@@ -1,47 +1,30 @@
 import { TestConfig } from "../types";
 
 const configs: TestConfig[] = [
-  // Deactivate the test since the query timeouts right now 26.04.2022
-  // {
-  //   chartId: "nQckT-nFEQe4",
-  //   name: "Column - Red list",
-  //   slug: "column-red-list",
-  // },
+  {
+    chartId: "nI6X4V_EcK2c",
+    name: "Bathing water quality (interactive filters hierarchie)",
+    slug: "bathing-water-quality-hierarchie",
+  },
   {
     chartId: "rOAYJqoZQKlU",
     name: "Column - Heavy metals",
     slug: "column-heavy-metals",
   },
   {
-    chartId: "cfNkIaMvN_xL",
-    name: "Scatterplot - Palmer Penguins",
-    slug: "scatterplot-palmer-penguins",
+    chartId: "jky5IEw6poT3",
+    name: "Map - NFI",
+    slug: "map-nfi",
   },
-  // {
-  //   chartId: "clQmsblRxpK7",
-  //   name: "Column - Waldfläsche (standard error bars)",
-  //   slug: "column-waldflasche",
-  // },
-  // {
-  //   chartId: "I0KucujlOLgb",
-  //   name: "Line - State accounts (segmented)",
-  //   slug: "line-state-accounts",
-  // },
   {
     chartId: "Z6Re21LHbOoP",
     name: "Pie - Red list",
     slug: "pie-red-list",
   },
-  // FIX: currently broken and should be rebuilt from PROD datasets and stored in local config
-  // {
-  //   chartId: "jky5IEw6poT3",
-  //   name: "Map - NFI: Topics by tree status",
-  //   slug: "map-nfi",
-  // },
   {
-    chartId: "nI6X4V_EcK2c",
-    name: "Bathing water quality (interactive filters hierarchie)",
-    slug: "bathing-water-quality-hierarchie",
+    chartId: "cfNkIaMvN_xL",
+    name: "Scatterplot - Palmer Penguins",
+    slug: "scatterplot-palmer-penguins",
   },
 ];
