--- app/test/__fixtures/config/int/line-state-accounts.json
@@ -1,56 +1,0 @@
-{
-  "key": "I0KucujlOLgb",
-  "data": {
-    "state": "DESCRIBING_CHART",
-    "dataSet": "https://culture.ld.admin.ch/sfa/StateAccounts_Office/1",
-    "dataSource": {
-      "type": "sparql",
-      "url": "https://lindas-cached.int.cluster.ldbar.ch/query"
-    },
-    "meta": {
-      "title": {
-        "de": "",
-        "fr": "",
-        "it": "",
-        "en": "Expenditure vs Revenue for Getreideverwaltung office"
-      },
-      "description": { "de": "", "fr": "", "it": "", "en": "" }
-    },
-    "chartConfig": {
-      "version": "1.2.1",
-      "chartType": "line",
-      "filters": {
-        "https://culture.ld.admin.ch/sfa/StateAccounts_Office/office": {
-          "type": "single",
-          "value": "https://culture.ld.admin.ch/sfa/StateAccounts_Office/Office/O1"
-        }
-      },
-      "interactiveFiltersConfig": {
-        "legend": { "active": false, "componentIri": "" },
-        "timeRange": {
-          "active": false,
-          "componentIri": "",
-          "presets": { "type": "range", "from": "", "to": "" }
-        },
-        "dataFilters": { "active": false, "componentIris": [] }
-      },
-      "fields": {
-        "x": { "componentIri": "http://www.w3.org/2006/time#Year" },
-        "y": { "componentIri": "http://schema.org/amount" },
-        "segment": {
-          "componentIri": "https://culture.ld.admin.ch/sfa/StateAccounts_Office/operationcharacter",
-          "palette": "category10",
-          "sorting": {
-            "sortingType": "byDimensionLabel",
-            "sortingOrder": "asc"
-          },
-          "colorMapping": {
-            "https://culture.ld.admin.ch/sfa/StateAccounts_Office/OperationCharacter/OC1": "#1f77b4",
-            "https://culture.ld.admin.ch/sfa/StateAccounts_Office/OperationCharacter/OC2": "#ff7f0e"
-          }
-        }
-      }
-    },
-    "activeField": "title"
-  }
-}
