+++ app/test/__fixtures/config/prod/configs.ts
@@ -1,21 +1,6 @@
 import { TestConfig } from "@/test/__fixtures/config/types";
 
 const configs: TestConfig[] = [
-  {
-    chartId: "column-traffic-pollution",
-    slug: "column-traffic-pollution",
-    name: "Column - Traffic pollution ",
-  },
-  {
-    chartId: "line-waldflasche",
-    slug: "line-waldflasche",
-    name: "Line Waldflasche",
-  },
-  {
-    chartId: "line-1",
-    slug: "line-1",
-    name: "Line 1 ",
-  },
   { chartId: "column-1", slug: "column-1", name: "Column 1" },
   { chartId: "column-2", slug: "column-2", name: "Column 2" },
   { chartId: "column-3", slug: "column-3", name: "Column 3" },
@@ -23,8 +8,18 @@ const configs: TestConfig[] = [
   { chartId: "column-5", slug: "column-5", name: "Column 5" },
   { chartId: "column-6", slug: "column-6", name: "Column 6" },
   { chartId: "column-7", slug: "column-7", name: "Column 7" },
+  {
+    chartId: "column-traffic-pollution",
+    slug: "column-traffic-pollution",
+    name: "Column - Traffic pollution ",
+  },
   { chartId: "line-1", slug: "line-1", name: "Line 1" },
   { chartId: "line-2", slug: "line-2", name: "Line 2" },
+  {
+    chartId: "line-waldflasche",
+    slug: "line-waldflasche",
+    name: "Line Waldflasche",
+  },
   { chartId: "pie-1", slug: "pie-1", name: "Pie 1" },
 ];
