+++ app/test/__fixtures/config/test/chartConfig-column-covid19.json
