+++ app/test/__fixtures/config/test/chartConfig-photovoltaik-und-gebaudeprogramm.json
