+++ app/test/__fixtures/config/test/chartConfig-table-covid19.json
